backend/requirements.txt
fastapi == 0.115.12
uvicorn == 0.34.2
sentence-transformers == 4.1.0 
torch == 2.7.0 
pandas == 2.2.3
pyarrow == 20.0.0
faiss-cpu == 1.11.0
tf-keras == 2.19.0 